requirements.txt
Django==2.1.5
gunicorn==19.9.0
django-heroku==0.3.1
django-bootstrap4==0.0.7
django-allauth==0.38.0
django-crispy-forms==1.7.2
social-auth-app-django==3.1.0
whitenoise==4.1.2
dj-database-url==0.5.0
python-dotenv==0.10.1
boto3==1.4.4
django-storages==1.7.1
pillow==5.4.1
django-autocomplete-light==3.3.4
django-filter==2.1.0
django-widget-tweaks==1.4.3